+++ tests/testthat/test-derive_vars_dtm.R
@@ -270,10 +270,23 @@ test_that("impute_dtc_dtm Test 7: impute month and day to mid, time to first", {
   )
 })
 
-## Test 8: min_dates parameter works ----
-test_that("impute_dtc_dtm Test 8: min_dates parameter works", {
+## Test 8: dtc is empty ----
+test_that("impute_dtc_dtm Test 8: dtc is empty", {
+  empty_impute <- impute_dtc_dtm(
+    dtc = character()
+  )
+
   expect_equal(
-    impute_dtc_dtm(c("2020-12", "2020-11", NA_character_),
+    empty_impute,
+    character(0)
+  )
+})
+
+## Test 9: min_dates argument works ----
+test_that("impute_dtc_dtm Test 9: min_dates argument works", {
+  expect_equal(
+    impute_dtc_dtm(
+      c("2020-12", "2020-11", NA_character_),
       min_dates = list(
         c(
           ymd_hms("2020-12-06T12:12:12"),
@@ -293,10 +306,11 @@ test_that("impute_dtc_dtm Test 8: min_dates parameter works", {
   )
 })
 
-## Test 9: max_dates parameter works ----
-test_that("impute_dtc_dtm Test 9: max_dates parameter works", {
+## Test 10: max_dates argument works ----
+test_that("impute_dtc_dtm Test 10: max_dates argument works", {
   expect_equal(
-    impute_dtc_dtm(c("2020-12", "2020-11", NA_character_, "2020-02-02"),
+    impute_dtc_dtm(
+      c("2020-12", "2020-11", NA_character_, "2020-02-02"),
       max_dates = list(
         c(ymd_hms("2020-12-06T12:12:12"), NA, ymd_hms("2020-09-13T08:30:00"), NA),
         c(ymd(""), ymd("2020-11-11"), ymd(""), ymd("2020-02-02"))
@@ -309,14 +323,158 @@ test_that("impute_dtc_dtm Test 9: max_dates parameter works", {
   )
 })
 
-## Test 10: min_dates length mismatch provides error ----
-test_that("impute_dtc_dtm Test 10: min_dates length mismatch provides error", {
+## Test 11: min_dates_strict argument only works ----
+test_that("impute_dtc_dtm Test 11: min_dates_strict argument only works", {
+  expect_equal(
+    impute_dtc_dtm(
+      c("2020-12", "2020-11", NA_character_),
+      min_dates_strict = list(
+        c(
+          ymd_hms("2020-12-06T12:12:12"),
+          NA,
+          NA
+        ),
+        c(
+          ymd_hms("2020-11-11T11:11:11"),
+          ymd_hms("2020-11-11T11:11:11"),
+          ymd_hms("2020-11-11T11:11:11")
+        )
+      ),
+      highest_imputation = "Y",
+      date_imputation = "first"
+    ),
+    c("2020-12-06T12:12:12", "2020-11-11T11:11:11", "2020-11-11T11:11:11")
+  )
+})
+
+## Test 12: min_dates and min_dates_strict arguments work ----
+test_that("impute_dtc_dtm Test 12: min_dates and min_dates_strict arguments work", {
+  expect_equal(
+    impute_dtc_dtm(
+      c("2020-12", "2020-11", NA_character_),
+      min_dates = list(
+        c(
+          ymd_hms("2020-12-06T12:12:12"),
+          NA,
+          NA
+        )
+      ),
+      min_dates_strict = list(
+        c(
+          ymd_hms("2020-11-11T11:11:11"),
+          ymd_hms("2020-11-11T11:11:11"),
+          ymd_hms("2020-11-11T11:11:11")
+        )
+      ),
+      highest_imputation = "Y",
+      date_imputation = "first"
+    ),
+    c("2020-12-06T12:12:12", "2020-11-11T11:11:11", "2020-11-11T11:11:11")
+  )
+})
+
+## Test 13: max_dates_strict argument only works ----
+test_that("impute_dtc_dtm Test 13: max_dates_strict argument only works", {
+  expect_equal(
+    impute_dtc_dtm(
+      c("2020-12", "2020-11", NA_character_, "2020-02-02"),
+      max_dates_strict = list(
+        c(ymd_hms("2020-12-06T12:12:12"), NA, ymd_hms("2020-09-13T08:30:00"), NA),
+        c(ymd(""), ymd("2020-11-11"), ymd(""), ymd("2020-02-02"))
+      ),
+      highest_imputation = "Y",
+      date_imputation = "last",
+      time_imputation = "last"
+    ),
+    c("2020-12-06T12:12:12", "2020-11-11T23:59:59", "2020-09-13T08:30:00", "2020-02-02T23:59:59")
+  )
+})
+
+## Test 14: max_dates and max_dates_strict arguments work ----
+test_that("impute_dtc_dtm Test 14: max_dates and max_dates_strict arguments work", {
+  expect_equal(
+    impute_dtc_dtm(
+      c("2020-12", "2020-11", NA_character_, "2020-02-02"),
+      max_dates_strict = list(
+        c(ymd_hms("2020-12-06T12:12:12"), NA, ymd_hms("2020-09-13T08:30:00"), NA)
+      ),
+      max_dates = list(
+        c(ymd(""), ymd("2020-11-11"), ymd(""), ymd("2020-02-02"))
+      ),
+      highest_imputation = "Y",
+      date_imputation = "last",
+      time_imputation = "last"
+    ),
+    c("2020-12-06T12:12:12", "2020-11-11T23:59:59", "2020-09-13T08:30:00", "2020-02-02T23:59:59")
+  )
+})
+
+## Test 15: max_dates_strict argument works ----
+test_that("impute_dtc_dtm Test 15: max_dates_strict argument works", {
+  expect_equal(
+    impute_dtc_dtm(
+      c("2020-12", "2020", "2020-11", NA_character_),
+      min_dates = list(
+        c(
+          ymd_hms("2020-12-06T12:12:12"),
+          ymd_hms("2020-12-06T12:12:12"),
+          ymd_hms("2020-12-06T12:12:12"),
+          NA
+        )
+      ),
+      max_dates_strict = list(
+        c(
+          ymd_hms("2020-12-05T00:00:00"),
+          ymd_hms("2020-12-15T00:00:00"),
+          ymd_hms("2020-11-12T12:00:00"),
+          NA
+        )
+      ),
+      highest_imputation = "Y",
+      date_imputation = "first",
+      time_imputation = "first"
+    ),
+    c("2020-12-01T00:00:00", "2020-12-06T12:12:12", "2020-11-01T00:00:00", "0000-01-01T00:00:00")
+  )
+})
+
+## Test 16: min_dates_strict argument works ----
+test_that("impute_dtc_dtm Test 16: min_dates_strict argument works", {
+  expect_equal(
+    impute_dtc_dtm(
+      c("2020-12", "2020", "2020-11", NA_character_),
+      max_dates = list(
+        c(
+          ymd_hms("2020-12-26T12:12:12"),
+          ymd_hms("2020-12-26T12:12:12"),
+          ymd_hms("2020-12-26T12:12:12"),
+          NA
+        )
+      ),
+      min_dates_strict = list(
+        c(
+          ymd_hms("2020-12-28T00:00:00"),
+          ymd_hms("2020-12-15T00:00:00"),
+          ymd_hms("2020-11-12T12:00:00"),
+          NA
+        )
+      ),
+      highest_imputation = "Y",
+      date_imputation = "last",
+      time_imputation = "last"
+    ),
+    c("2020-12-31T23:59:59", "2020-12-26T12:12:12", "2020-11-30T23:59:59", "9999-12-31T23:59:59")
+  )
+})
+
+## Test 17: min_dates length mismatch provides error ----
+test_that("impute_dtc_dtm Test 17: min_dates length mismatch provides error", {
   expect_snapshot(
     impute_dtc_dtm(
       c("2020-12", NA_character_),
       min_dates = list(
-        c(ymd_hms("2020-12-06T12:12:12")),
-        c(ymd_hms("2020-11-11T11:11:11"))
+        ymd_hms("2020-12-06T12:12:12"),
+        ymd_hms("2020-11-11T11:11:11")
       ),
       highest_imputation = "Y"
     ),
@@ -324,14 +482,14 @@ test_that("impute_dtc_dtm Test 10: min_dates length mismatch provides error", {
   )
 })
 
-## Test 11: max_dates length mismatch provides error ----
-test_that("impute_dtc_dtm Test 11: max_dates length mismatch provides error", {
+## Test 18: max_dates length mismatch provides error ----
+test_that("impute_dtc_dtm Test 18: max_dates length mismatch provides error", {
   expect_snapshot(
     impute_dtc_dtm(
       c("2020-12", NA_character_),
       max_dates = list(
-        c(ymd_hms("2020-12-06T12:12:12")),
-        c(ymd_hms("2020-11-11T11:11:11"))
+        ymd_hms("2020-12-06T12:12:12"),
+        ymd_hms("2020-11-11T11:11:11")
       ),
       highest_imputation = "Y",
       date_imputation = "last"
@@ -340,8 +498,8 @@ test_that("impute_dtc_dtm Test 11: max_dates length mismatch provides error", {
   )
 })
 
-## Test 12: Error if null min/max_dates when highest_imputation = Y ----
-test_that("impute_dtc_dtm Test 12: Error if null min/max_dates when highest_imputation = Y", {
+## Test 19: Error if null min/max_dates when highest_imputation = Y ----
+test_that("impute_dtc_dtm Test 19: Error if null min/max_dates when highest_imputation = Y", {
   expect_snapshot(
     impute_dtc_dtm(
       c("2020-12", NA_character_),
@@ -351,8 +509,8 @@ test_that("impute_dtc_dtm Test 12: Error if null min/max_dates when highest_impu
   )
 })
 
-## Test 13: wrong input to `date_imputation` ----
-test_that("impute_dtc_dtm Test 13: wrong input to `date_imputation`", {
+## Test 20: wrong input to `date_imputation` ----
+test_that("impute_dtc_dtm Test 20: wrong input to `date_imputation`", {
   # wrong format
   expect_snapshot(
     impute_dtc_dtm(
@@ -419,8 +577,8 @@ test_that("impute_dtc_dtm Test 13: wrong input to `date_imputation`", {
 })
 
 # convert_dtc_to_dtm ----
-## Test 14: Convert a complete -- DTC into a date time object ----
-test_that("convert_dtc_to_dtm Test 14: Convert a complete -- DTC into a date time object", {
+## Test 21: Convert a complete -- DTC into a date time object ----
+test_that("convert_dtc_to_dtm Test 21: Convert a complete -- DTC into a date time object", {
   expect_equal(
     convert_dtc_to_dtm("2019-07-18T15:25:40.243"),
     ymd_hms("2019-07-18T15:25:40.243")
@@ -428,8 +586,8 @@ test_that("convert_dtc_to_dtm Test 14: Convert a complete -- DTC into a date tim
 })
 
 # compute_tmf ----
-## Test 15: compute TMF ----
-test_that("compute_tmf Test 15: compute TMF", {
+## Test 22: compute TMF ----
+test_that("compute_tmf Test 22: compute TMF", {
   input_dtc <- c(
     "2019-07-18T15:25:52",
     "2019-07-18T15:25",
@@ -489,8 +647,8 @@ test_that("compute_tmf Test 15: compute TMF", {
   )
 })
 
-## Test 16: throws ERROR when ignore_seconds_flag  = T (default) and seconds are present ----
-test_that("compute_tmf Test 16: throws ERROR when ignore_seconds_flag  = T and seconds are present", { # nolint
+## Test 23: throws ERROR when ignore_seconds_flag  = T and seconds are present ----
+test_that("compute_tmf Test 23: throws ERROR when ignore_seconds_flag  = T and seconds are present", { # nolint
   expect_snapshot(
     compute_tmf(
       dtc = c("2020-11-11T11:11:11", "2020-11-11T11:11"),
@@ -502,8 +660,8 @@ test_that("compute_tmf Test 16: throws ERROR when ignore_seconds_flag  = T and s
   )
 })
 
-## Test 17: ignore_seconds_flag  = TRUE ----
-test_that("compute_tmf Test 17: ignore_seconds_flag  = TRUE", {
+## Test 24: ignore_seconds_flag  = TRUE ----
+test_that("compute_tmf Test 24: ignore_seconds_flag  = TRUE", {
   expect_equal(
     compute_tmf(
       dtc = c("2020-11-11T11:11", "2020-11-11T11"),
@@ -518,8 +676,8 @@ test_that("compute_tmf Test 17: ignore_seconds_flag  = TRUE", {
 
 # derive_vars_dtm ----
 
-## Test 18: default behavior ----
-test_that("derive_vars_dtm Test 18: default behavior", {
+## Test 25: default behavior ----
+test_that("derive_vars_dtm Test 25: default behavior", {
   input <- tibble::tribble(
     ~XXSTDTC,
     "2019-07-18T15:25:40",
@@ -556,8 +714,8 @@ test_that("derive_vars_dtm Test 18: default behavior", {
   )
 })
 
-## Test 19: date imputed to first, auto DTF/TMF ----
-test_that("derive_vars_dtm Test 19: date imputed to first, auto DTF/TMF", {
+## Test 26: date imputed to first, auto DTF/TMF ----
+test_that("derive_vars_dtm Test 26: date imputed to first, auto DTF/TMF", {
   input <- tibble::tribble(
     ~XXSTDTC,
     "2019-07-18T15:25:40",
@@ -595,8 +753,8 @@ test_that("derive_vars_dtm Test 19: date imputed to first, auto DTF/TMF", {
   )
 })
 
-## Test 20: date and time imputed to last, no DTF/TMF ----
-test_that("derive_vars_dtm Test 20: date and time imputed to last, no DTF/TMF", {
+## Test 27: date and time imputed to last, no DTF/TMF ----
+test_that("derive_vars_dtm Test 27: date and time imputed to last, no DTF/TMF", {
   input <- tibble::tribble(
     ~XXSTDTC,
     "2019-07-18T15:25:40",
@@ -635,8 +793,8 @@ test_that("derive_vars_dtm Test 20: date and time imputed to last, no DTF/TMF", 
   )
 })
 
-## Test 21: date and time imputed to last, DTF only ----
-test_that("derive_vars_dtm Test 21: date and time imputed to last, DTF only", {
+## Test 28: date and time imputed to last, DTF only ----
+test_that("derive_vars_dtm Test 28: date and time imputed to last, DTF only", {
   input <- tibble::tribble(
     ~XXSTDTC,
     "2019-07-18T15:25:40",
@@ -675,8 +833,8 @@ test_that("derive_vars_dtm Test 21: date and time imputed to last, DTF only", {
   )
 })
 
-## Test 22: date imputed to MID, time to first, TMF only ----
-test_that("derive_vars_dtm Test 22: date imputed to MID, time to first, TMF only", {
+## Test 29: date imputed to MID, time to first, TMF only ----
+test_that("derive_vars_dtm Test 29: date imputed to MID, time to first, TMF only", {
   input <- tibble::tribble(
     ~XXSTDTC,
     "2019-07-18T15:25:40",
@@ -716,8 +874,8 @@ test_that("derive_vars_dtm Test 22: date imputed to MID, time to first, TMF only
   )
 })
 
-## Test 23: No re-derivation is done if --DTF variable already exists ----
-test_that("derive_vars_dtm Test 23: No re-derivation is done if --DTF variable already exists", {
+## Test 30: No re-derivation is done if --DTF variable already exists ----
+test_that("derive_vars_dtm Test 30: No re-derivation is done if --DTF variable already exists", {
   input <- tibble::tribble(
     ~XXSTDTC,
     "2019-07-18T15:25:40",
@@ -758,8 +916,8 @@ test_that("derive_vars_dtm Test 23: No re-derivation is done if --DTF variable a
   )
 })
 
-## Test 24: max_dates parameter works as expected ----
-test_that("derive_vars_dtm Test 24: max_dates parameter works as expected", {
+## Test 31: max_dates argument works as expected ----
+test_that("derive_vars_dtm Test 31: max_dates argument works as expected", {
   expected_output <- tibble::tribble(
     ~XXSTDTC,    ~ASTDTM,                        ~ASTDTF, ~ASTTMF,
     "2019-02",   ymd_hms("2019-02-10T00:00:00"), "D",     "H",
@@ -784,8 +942,8 @@ test_that("derive_vars_dtm Test 24: max_dates parameter works as expected", {
   )
 })
 
-## Test 25: NA imputation for highest_imputation = Y & max_dates ----
-test_that("derive_vars_dtm Test 25: NA imputation for highest_imputation = Y & max_dates", {
+## Test 32: NA imputation for highest_imputation = Y & max_dates ----
+test_that("derive_vars_dtm Test 32: NA imputation for highest_imputation = Y & max_dates", {
   actual <- data.frame(
     AESTDTC = c(NA_character_, NA_character_),
     TRTSDTM = c(ymd_hms("2022-01-01 23:59:59"), NA)
@@ -812,8 +970,8 @@ test_that("derive_vars_dtm Test 25: NA imputation for highest_imputation = Y & m
   expect_dfs_equal(actual, expected, keys = c("ASTDTM", "ASTDTF", "ASTTMF"))
 })
 
-## Test 26: Error for highest_imputation = Y & max_dates but date_imputation = first ----
-test_that("derive_vars_dtm Test 26: Error for highest_imputation = Y & max_dates but date_imputation = first", { # nolint
+## Test 33: Error for highest_imputation = Y & max_dates but date_imputation = first ----
+test_that("derive_vars_dtm Test 33: Error for highest_imputation = Y & max_dates but date_imputation = first", { # nolint
   expect_snapshot(
     (data.frame(
       AESTDTC = c(NA_character_, NA_character_),
@@ -833,8 +991,8 @@ test_that("derive_vars_dtm Test 26: Error for highest_imputation = Y & max_dates
   )
 })
 
-## Test 27: NA imputation for highest_imputation = Y & min_dates ----
-test_that("derive_vars_dtm Test 27: NA imputation for highest_imputation = Y & min_dates", {
+## Test 34: NA imputation for highest_imputation = Y & min_dates ----
+test_that("derive_vars_dtm Test 34: NA imputation for highest_imputation = Y & min_dates", {
   actual <- data.frame(
     AESTDTC = c(NA_character_, NA_character_),
     TRTSDTM = c(ymd_hms("2022-01-01 23:59:59"), NA)
@@ -861,8 +1019,8 @@ test_that("derive_vars_dtm Test 27: NA imputation for highest_imputation = Y & m
   expect_dfs_equal(actual, expected, keys = c("ASTDTM", "ASTDTF", "ASTTMF"))
 })
 
-## Test 28: Error for highest_imputation = Y & min_dates but date_imputation = last ----
-test_that("derive_vars_dtm Test 28: Error for highest_imputation = Y & min_dates but date_imputation = last", { # nolint
+## Test 35: Error for highest_imputation = Y & min_dates but date_imputation = last ----
+test_that("derive_vars_dtm Test 35: Error for highest_imputation = Y & min_dates but date_imputation = last", { # nolint
   expect_snapshot(
     data.frame(
       AESTDTC = c(NA_character_, NA_character_),
@@ -882,8 +1040,8 @@ test_that("derive_vars_dtm Test 28: Error for highest_imputation = Y & min_dates
   )
 })
 
-## Test 29: NA imputation for highest_imputation = Y but null min/max dates fails ----
-test_that("derive_vars_dtm Test 29: NA imputation for highest_imputation = Y but null min/max dates fails", { # nolint
+## Test 36: NA imputation for highest_imputation = Y but null min/max dates fails ----
+test_that("derive_vars_dtm Test 36: NA imputation for highest_imputation = Y but null min/max dates fails", { # nolint
   expect_snapshot(
     data.frame(
       AESTDTC = c(NA_character_, NA_character_),
@@ -902,8 +1060,8 @@ test_that("derive_vars_dtm Test 29: NA imputation for highest_imputation = Y but
   )
 })
 
-## Test 30: Supplying both min/max dates for highest_imputation = Y works ----
-test_that("derive_vars_dtm Test 30: Supplying both min/max dates for highest_imputation = Y works", { # nolint
+## Test 37: Supplying both min/max dates for highest_imputation = Y works ----
+test_that("derive_vars_dtm Test 37: Supplying both min/max dates for highest_imputation = Y works", { # nolint
   actual <- data.frame(
     AESTDTC = c(NA_character_, NA_character_),
     TRTSDTM = c(ymd_hms("2022-01-01 23:59:59"), NA),
@@ -932,8 +1090,8 @@ test_that("derive_vars_dtm Test 30: Supplying both min/max dates for highest_imp
   expect_dfs_equal(actual, expected, keys = c("ASTDTM", "ASTDTF", "ASTTMF"))
 })
 
-## Test 31: Supplying both min/max dates for highest_imputation = Y works ----
-test_that("derive_vars_dtm Test 31: Supplying both min/max dates for highest_imputation = Y works", { # nolint
+## Test 38: Supplying both min/max dates for highest_imputation = Y works ----
+test_that("derive_vars_dtm Test 38: Supplying both min/max dates for highest_imputation = Y works", { # nolint
   actual <- data.frame(
     AESTDTC = c(NA_character_, NA_character_),
     TRTSDTM = c(ymd_hms("2022-01-01 23:59:59"), NA),
@@ -962,8 +1120,8 @@ test_that("derive_vars_dtm Test 31: Supplying both min/max dates for highest_imp
   expect_dfs_equal(actual, expected, keys = c("AENDTM", "AENDTF", "AENTMF"))
 })
 
-## Test 32: catch ignore_seconds_flag error ----
-test_that("derive_vars_dtm Test 32: catch ignore_seconds_flag error", {
+## Test 39: catch ignore_seconds_flag error ----
+test_that("derive_vars_dtm Test 39: catch ignore_seconds_flag error", {
   input <- tibble::tribble(
     ~XXSTDTC,
     "2019-07-18T15:25:40",
@@ -983,17 +1141,5 @@ test_that("derive_vars_dtm Test 32: catch ignore_seconds_flag error", {
       ignore_seconds_flag = TRUE
     ),
     error = TRUE
-  )
-})
-
-# Test 33: impute_dtc_dtm returns an empty character vector where dtc is empty
-test_that("derive_vars_dt Test 30: impute_dtc_dt where dtc is empty", {
-  empty_impute <- impute_dtc_dtm(
-    dtc = character()
-  )
-
-  expect_equal(
-    empty_impute,
-    character(0)
   )
 })