+++ NEWS.md
@@ -4,13 +4,32 @@
 
 ## Updates of Existing Functions
 
+- The `min_dates_strict` and `max_dates_strict` arguments were added to
+`derive_vars_dt()`, `derive_vars_dtm()`, `convert_dtc_to_dt()`,
+`convert_dtc_to_dtm()`, `impute_dtc_dt()`, `impute_dtc_dtm()`. The new arguments
+can be used to avoid invalid imputed dates. E.g., if `AESTDTC` is imputed with
+`min_dates = exprs(TRTSDTM)`, `max_dates_strict = exprs(AENDTM)` can be
+specified to ensure the imputed event start date is not after the event end
+date (the `ADAE` template and the "Creating an OCCDS ADaM" vignette were updated
+accordingly).
+
+    See the `derive_vars_dt()` documentation and the "Avoid Imputed Dates Before
+    a Particular Date" section of the "Date and Time Imputation" vignette
+    (`vignette("imputation")`) for details and examples. (#2708)
+    
+- `derive_vars_merged_lookup()` was enhanced with a new `check_not_mapped_type` argument
+which allows users to control whether a message, warning, error, or no message at all
+is issued when some records are not mapped using the lookup dataset. The `print_not_mapped`
+argument will thus be deprecated. (#3188)
+
 ## Breaking Changes
 
 - The following functions are entering the next phase of the [deprecation process](https://pharmaverse.github.io/admiraldev/articles/programming_strategy.html#deprecation):
 
   **Phase 1 (message)**
 
-  No functions or arguments in this phase.
+  - `derive_vars_merged_lookup(print_not_mapped = )` is deprecated in favor of the 
+  `check_not_mapped_type` argument.
 
   **Phase 2 (warning)**
 
@@ -33,6 +52,8 @@
   - `call_user_fun()` has been removed.
 
 ## Documentation
+
+- Corrected the ADLB template label for `PARAMCD = "CK"` from `"Creatinine Kinase"` to `"Creatine Kinase"` (#3170).
 
 - Updated the "Lab Grading" vignette to replace the deprecated `abnormal_indicator` argument reference
   with the new `high_indicator` argument in `derive_var_atoxgr_dir()`.